backend/pyproject.toml
[project]
name = "linkedin-scraper-resume"
version = "1.0.0"
description = "Automated LinkedIn job scraper with AI-powered CV tailoring using Google Gemini"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "authlib>=1.6.5",
    "bcrypt>=5.0.0",
    "cryptography>=46.0.3",
    "en-core-web-sm",
    "fastapi>=0.121.3",
    "firebase-admin>=7.1.0",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "markdown>=3.10",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "passlib>=1.7.4",
    "playwright>=1.55.0",
    "plotly>=6.5.0",

    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.7.2",
    "spacy>=3.8.11",
    "supabase>=2.24.0",
    "uvicorn>=0.38.0",
    "weasyprint>=66.0",
]

[project.urls]
Homepage = "https://github.com/khaldon/linkedin-scrapper-resume"
Repository = "https://github.com/khaldon/linkedin-scrapper-resume.git"
Issues = "https://github.com/khaldon/linkedin-scrapper-resume/issues"

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }
